requirements.txt
asgiref==3.4.0
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.28.0
dj-database-url==0.5.0
Django==3.2.5
django-bootstrap4==21.1
django-decouple==2.1
django-multiselectfield==0.1.12
django-on-heroku==1.1.2
django-tinymce==3.3.0
idna==3.2
psycopg2-binary==2.9.2
python-dateutil==1.5
python-decouple==3.5
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.26.6
whitenoise==5.3.0
